01_openai_agents/tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.4.400",
    "openai-agents>=0.0.8",
    "python-dotenv>=1.1.0",
]
